+++ src/Entity/Page.php
@@ -78,7 +78,7 @@ class Page implements HasBlocksInterface, \Stringable
 
     #[ORM\ManyToMany(targetEntity: Block::class, cascade: ['persist', 'remove'])]
     #[ORM\JoinTable(name: 'site_page_blocks')]
-    #[ORM\OrderBy(['position' => 'ASC'])]
+    #[ORM\OrderBy(['position' => \SortDirection::Ascending])]
     private Collection $blocks;
 
     #[ORM\Column(type: Types::BOOLEAN, options: ['default' => false])]